Add files via upload
"DiD yOu AcTuAlLY JuSt UnFolLoW Me??? Sam, please don't unfollow me. Sam this is serious."

The five stages, denial, anger, bargaining, depression and acceptance are a part of the framework that makes up our learning to live with the one we lost. They are tools to help us frame and identify what we may be feeling. But they are not stops on some linear timeline in grief.